refactor(adapter-node): extract HTTP listen helper into shared lib
Verified: precommit-verify △ (99816e0483b7c77d)
