+++ packages/adapter-node/src/on-node.ts
@@ -1,5 +1,3 @@
-import type { ServerType } from '@hono/node-server';
-import { serve } from '@hono/node-server';
 import type {
   ConfigClass,
   ConfiguredFeature,
@@ -9,18 +7,12 @@ import type {
 } from '@zeltjs/core';
 import { HttpFeature } from '@zeltjs/core';
 
+import type { ListenOptions, ServerHandle } from './listen.lib';
+import { createListenForHttp } from './listen.lib';
 import { NodeCliConfig } from './node-cli.config';
 import { ProcessEnvAdaptor } from './process-env.adaptor';
 
-type ListenOptions = {
-  readonly port?: number;
-  readonly hostname?: string;
-};
-
-export type ServerHandle = {
-  readonly address: { port: number; address: string };
-  readonly shutdown: () => Promise<void>;
-};
+export type { ServerHandle } from './listen.lib';
 
 export type NodeAppOptions = {
   readonly configs?: readonly ConfigClass<object>[];
@@ -51,36 +43,6 @@ export type NodeApp = RuntimeApp<readonly ConfiguredFeature[]> & EnvironmentNode
 type NodeAppForFeatures<F extends readonly ConfiguredFeature[]> = RuntimeApp<F> &
   EnvironmentNodeAppPart &
   NodeNamespacedCapabilities<F>;
-
-const closeServer = (server: ServerType): Promise<void> =>
-  new Promise<void>((resolve, reject) => {
-    server.close((err) => (err ? reject(err) : resolve()));
-  });
-
-const createListenForHttp = (
-  appFetch: (request: Request) => Promise<Response>,
-  registerShutdown: (callback: () => Promise<void>) => () => Promise<void>,
-): ((portOrOptions?: number | ListenOptions) => Promise<ServerHandle>) => {
-  return async (portOrOptions?: number | ListenOptions): Promise<ServerHandle> => {
-    const listenOptions: ListenOptions =
-      typeof portOrOptions === 'number' ? { port: portOrOptions } : (portOrOptions ?? {});
-
-    const port = listenOptions.port ?? 3000;
-    const hostname = listenOptions.hostname ?? '0.0.0.0';
-
-    let server!: ServerType;
-    const serverReady = new Promise<{ port: number; address: string }>((resolve) => {
-      server = serve({ fetch: appFetch, port, hostname }, (info) =>
-        resolve({ port: info.port, address: info.address }),
-      );
-    });
-    const shutdown = registerShutdown(() => closeServer(server));
-
-    const address = await serverReady;
-
-    return { address, shutdown };
-  };
-};
 
 const getArgs = (): readonly string[] => globalThis.process.argv.slice(2);
